backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
watchdog==5.0.2
orjson==3.10.7
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-logging==0.47b0

